+++ Source/MessageExchangeDefinition/gdcmCStoreMessages.cxx
@@ -187,7 +187,7 @@ const DataSet* inDataSet = &file.GetDataSet();
 static uint32_t messageid = 1;
   {
   Attribute<0x0,0x110> at = { 0 };
-  at.SetValue( messageid++ );
+  at.SetValue( (unsigned short)messageid++ );
   assert( messageid < std::numeric_limits<uint32_t>::max());
   ds.Insert( at.GetAsDataElement() );
   }
@@ -305,7 +305,7 @@ std::vector<PresentationDataValue> CStoreRSP::ConstructPDV(const DataSet* inData
     {
     // Message ID Being Responded To
     Attribute<0x0,0x120> at = { 1 };
-    at.SetValue( theMessageID );
+    at.SetValue( (unsigned short)theMessageID );
     ds.Insert( at.GetAsDataElement() );
     }
     {
