+++ Testing/Source/MessageExchangeDefinition/Cxx/TestSCUFunctions.cxx
@@ -106,8 +106,8 @@ int TestSCUFunctions(int argc, char *argv[])
   gdcm::Trace::WarningOff();
   std::string aetitle = argv[1]; // the ae title of this computer
   std::string call = argv[2]; // the ae title of the server
-  int portno = atoi(argv[3]); // the port of the server
-  int moveReturnPort = atoi(argv[4]); // the port over which return cstore scps are done for cmove
+  uint16_t portno = (uint16_t)atoi(argv[3]); // the port of the server
+  uint16_t moveReturnPort = (uint16_t)atoi(argv[4]); // the port over which return cstore scps are done for cmove
   std::string remote = argv[5]; //the ip address of the remote server
   std::string tmpdir = gdcm::Testing::GetTempDirectory( "TestSCUFunctions" );
   std::string outputDir = tmpdir; //place to where data is returned by cmove
